+++ packages/app-store/_utils/payments/getPaymentAppData.ts
@@ -1,6 +1,5 @@
 import type { z } from "zod";
 
-import type { BookerEvent } from "@calcom/features/bookings/types";
 import type { EventTypeMetaDataSchema } from "@calcom/prisma/zod-utils";
 
 import type { appDataSchemas } from "../../apps.schemas.generated";
@@ -10,7 +9,9 @@ import { eventTypeMetaDataSchemaWithTypedApps } from "../../zod-utils";
 import { getEventTypeAppData } from "../getEventTypeAppData";
 
 export function getPaymentAppData(
-  _eventType: Pick<BookerEvent, "price" | "currency"> & {
+  _eventType: {
+    price: number;
+    currency: string;
     metadata: z.infer<typeof EventTypeMetaDataSchema>;
   },
   forcedGet?: boolean
