+++ packages/features/bookings/types.ts
@@ -1,11 +1,11 @@
 import type { ErrorOption, FieldPath } from "react-hook-form";
 
 import type { RegularBookingCreateResult } from "@calcom/features/bookings/lib/dto/types";
+import type { Slots } from "@calcom/features/calendars/lib/slots";
 import type { SchedulingType } from "@calcom/prisma/enums";
 import type { RouterOutputs } from "@calcom/trpc/react";
 import type { AppsStatus } from "@calcom/types/Calendar";
 
-import type { DatePickerProps } from "../calendars/DatePicker";
 import type { BookingCreateBody } from "./lib/bookingCreateBodySchema";
 
 export type PublicEvent = NonNullable<RouterOutputs["viewer"]["public"]["event"]>;
@@ -29,8 +29,8 @@ type BookerEventUser = Pick<
 
 type BookerEventProfile = Pick<PublicEvent["profile"], "name" | "image" | "bookerLayouts">;
 
-// marked as required to keep responsibility on consumers to handle the case where slots is undefined
-export type Slots = Required<NonNullable<DatePickerProps["slots"]>>;
+// Re-export Slots from the server-safe location
+export type { Slots };
 
 export type BookerEvent = Pick<
   PublicEvent,
