+++ packages/trpc/server/routers/viewer/eventTypes/heavy/create.schema.ts
@@ -1,6 +1,6 @@
 import type { z } from "zod";
 
-import { createEventTypeInput } from "@calcom/features/eventtypes/lib/types";
+import { createEventTypeInput } from "@calcom/features/eventtypes/lib/types.server";
 
 export const ZCreateInputSchema = createEventTypeInput;
