+++ packages/trpc/server/routers/viewer/eventTypes/heavy/duplicate.schema.ts
@@ -1,6 +1,6 @@
 import type { z } from "zod";
 
-import { EventTypeDuplicateInput } from "@calcom/features/eventtypes/lib/types";
+import { EventTypeDuplicateInput } from "@calcom/features/eventtypes/lib/types.server";
 
 export const ZDuplicateInputSchema = EventTypeDuplicateInput;
